[BE]:[chore] 좌표값에 따른 범위 계산 수정 필요

+++ Backend/finedUI/src/main/java/com/ssafy/finedUi/registInfo/update/RegistInfoUpdateController.java
@@ -21,7 +21,7 @@ public class RegistInfoUpdateController {
     }
 
     @PatchMapping
-    public ResponseEntity<Object> isMissingChange(@RequestParam Long registId, Integer longitude, Integer latitude) {
+    public ResponseEntity<Object> isMissingChange(@RequestParam Long registId, Double longitude, Double latitude) {
         return ResponseHandler.generateResponse(true, "CHANGE", HttpStatus.OK, registInfoUpdateService.isMissingChange(registId, longitude, latitude));
     }
 }
